+++ tests/test_ulmfit.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+import datetime
+import os
+import sys
+import unittest
+
+
+class TestUlmfitPackage(unittest.TestCase):
+
+    def setUp():
+        pass
